move useComments to global hooks dir
