+++ src/hooks/index.js
@@ -7,6 +7,7 @@ export { default as usePaywall } from "./api/usePaywall";
 export { default as usePolicy } from "./api/usePolicy";
 export { default as useProposalsBatch } from "./api/useProposalsBatch";
 export { default as useTimestamps } from "./api/useTimestamps";
+export { default as useComments } from "./api/useComments";
 export { default as useSubContractors } from "./api/useSubContractors";
 export { default as useSupervisors } from "./api/useSupervisors";
 export { default as useUserDetail } from "./api/useUserDetail";
